+++ apps/client/src/shared/api/uploadFileToPresignedUrl.ts
@@ -1,10 +1,13 @@
 import { HttpStatusError } from '@/shared/api/apiError'
 
-import type { ReviewImageUploadTarget } from '@/pages/reviewNew/api/issueReviewImageUploads'
+export interface PresignedUrlUploadTarget {
+  uploadUrl: string
+  uploadMethod: string
+}
 
 export const uploadFileToPresignedUrl = async (
   file: File,
-  { uploadMethod, uploadUrl }: ReviewImageUploadTarget,
+  { uploadMethod, uploadUrl }: PresignedUrlUploadTarget,
 ) => {
   const response = await fetch(uploadUrl, {
     method: uploadMethod,
